+++ druid-shell/src/platform/mac/window.rs
@@ -53,7 +53,7 @@ use super::util::{assert_main_thread, make_nsstring};
 use crate::common_util::IdleCallback;
 use crate::dialog::{FileDialogOptions, FileDialogType, FileInfo};
 use crate::keyboard_types::KeyState;
-use crate::mouse::{Cursor, MouseButton, MouseButtons, MouseEvent};
+use crate::mouse::{Cursor, CursorDesc, MouseButton, MouseButtons, MouseEvent};
 use crate::region::Region;
 use crate::scale::Scale;
 use crate::window::{IdleToken, TimerToken, WinHandler, WindowLevel, WindowState};
@@ -144,6 +144,10 @@ struct ViewState {
     keyboard_state: KeyboardState,
     text: PietText,
 }
+
+#[derive(Clone)]
+// TODO: support custom cursors
+pub struct CustomCursor;
 
 impl WindowBuilder {
     pub fn new(_app: Application) -> WindowBuilder {
@@ -888,9 +892,16 @@ impl WindowHandle {
                 Cursor::NotAllowed => msg_send![nscursor, operationNotAllowedCursor],
                 Cursor::ResizeLeftRight => msg_send![nscursor, resizeLeftRightCursor],
                 Cursor::ResizeUpDown => msg_send![nscursor, resizeUpDownCursor],
+                // TODO: support custom cursors
+                Cursor::Custom(_) => msg_send![nscursor, arrowCursor],
             };
             let () = msg_send![cursor, set];
         }
+    }
+
+    pub fn make_cursor(&self, _cursor_desc: &CursorDesc) -> Option<Cursor> {
+        log::warn!("Custom cursors are not yet supported in the macOS backend");
+        None
     }
 
     pub fn request_timer(&self, deadline: std::time::Instant) -> TimerToken {