+++ druid/src/lib.rs
@@ -177,7 +177,7 @@ pub use piet::{
 // these are the types from shell that we expose; others we only use internally.
 pub use shell::keyboard_types;
 pub use shell::{
-    Application, Clipboard, ClipboardFormat, Code, Cursor, Error as PlatformError,
+    Application, Clipboard, ClipboardFormat, Code, Cursor, CursorDesc, Error as PlatformError,
     FileDialogOptions, FileInfo, FileSpec, FormatId, HotKey, ImageBuf, KbKey, KeyEvent, Location,
     Modifiers, Monitor, MouseButton, MouseButtons, RawMods, Region, Scalable, Scale, Screen,
     SysMods, TimerToken, WindowHandle, WindowState,